requirements.txt
streamlit
pandas
scikit-learn
sentence-transformers
transformers
torch
python-dotenv
matplotlib         
textblob           
nltk 